+++ src/__tests__/About.test.tsx
@@ -1,6 +1,6 @@
 import { describe } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { AboutPage } from '../pages/AboutPage';
+import { AboutPage } from '../pages/About/AboutPage';
 
 describe('About page component', () => {
   test('renders the not found page', () => {
