+++ src/__tests__/Modal.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { ICharacter } from '../models';
 import { Modal } from '../components/Modal';
-import { Character } from '../components/Character';
+import { Character } from '../pages/Home/components/Character';
 
 describe('Modal component', () => {
   const testCharacter: ICharacter = {
