+++ src/__tests__/NotFoundPage.test.tsx
@@ -1,6 +1,6 @@
 import { describe } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { NotFoundPage } from '../pages/NotFoundPage';
+import { NotFoundPage } from '../pages/NotFound/NotFoundPage';
 
 describe('NotFoundPage component', () => {
   test('renders the not found page', () => {
