+++ src/components/Layout.tsx
@@ -1,9 +1,36 @@
-import { NavLink, Link, Outlet } from 'react-router-dom';
+import { NavLink, Link, Outlet, useLocation } from 'react-router-dom';
+import { useEffect, useState } from 'react';
 import '../styles/layout.scss';
-import { useLayout } from '../hooks/LayoutHook';
+
+enum PageName {
+  HOME = '/',
+  ABOUT = '/about',
+  FORM = '/form',
+}
 
 export function Layout(): JSX.Element {
-  const { currentPageName } = useLayout();
+  const location = useLocation();
+  const [currentPageName, setCurrentPageName] = useState<string>();
+
+  useEffect(() => {
+    setCurrentPageName(() => {
+      switch (location.pathname) {
+        case PageName.HOME:
+          return 'Home';
+        case PageName.ABOUT:
+          return 'About us';
+        case PageName.FORM:
+          return 'Form';
+        default:
+          return 'Page not found';
+      }
+    });
+  }, [location.pathname]);
+
+  useEffect(() => {
+    document.title = currentPageName + ' | The Rick and Morty';
+  }, [currentPageName]);
+
   return (
     <>
       <header className="header">
