+++ src/pages/Home/HomePage.tsx
@@ -1,8 +1,15 @@
-import { useEffect, useState } from 'react';
-import { ICharacter, IResponse } from '../models';
 import axios, { AxiosError } from 'axios';
+import { CharacterMini } from './components/CharacterMini';
+import { Search } from './components/Search';
+import { Loader } from '../../components/Loader';
+import { ErrorMessage } from '../../components/ErrorMessage';
+import { Modal } from '../../components/Modal';
+import { Character } from './components/Character';
+import { useEffect, useState } from 'react';
+import { ICharacter, IResponse } from '../../models';
+import '../../styles/homePage.scss';
 
-export const useHomePage = () => {
+export function HomePage(): JSX.Element {
   const [characters, setCharacters] = useState<ICharacter[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
   const [error, setError] = useState<string>('');
@@ -39,5 +46,27 @@ export const useHomePage = () => {
     return fetchData(searchValue);
   };
 
-  return { characters, loading, error, handleSearch, character, setCharacter };
-};
+  return (
+    <>
+      <main className="main-home">
+        <Search onSearch={handleSearch} />
+        {loading && <Loader />}
+        {error && <ErrorMessage error={error} />}
+        <section className="character">
+          {characters.map((character) => (
+            <CharacterMini
+              character={character}
+              key={character.id}
+              onClick={() => setCharacter(character)}
+            />
+          ))}
+        </section>
+      </main>
+      {character && (
+        <Modal onClose={() => setCharacter(null)}>
+          <Character character={character} />
+        </Modal>
+      )}
+    </>
+  );
+}
